so there is some protection against clearing the entire database

+++ src/utils/test.utils.ts
@@ -2,6 +2,10 @@ import { PrismaClient, Prisma } from "@prisma/client";
 const prisma = new PrismaClient();
 
 export async function clearEntireDatabase(): Promise<Prisma.BatchPayload[]> {
+  if (process.env.NODE_ENV != "test") {
+    throw Error("Only use this function in a testing environment");
+  }
+
   const results = [await prisma.lastfmListen.deleteMany()];
 
   results.push(await prisma.lastfmAccount.deleteMany());
